requirements.txt
asgiref==3.8.1
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
Django==5.2.1
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pycparser==2.22
pypdfium2==4.30.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
